move grounding dino from projects to mmdet package
