+++ mmdet/models/language_models/__init__.py
@@ -1,4 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .bert import BertModel
+from .grounding_dino_bert import GroundingDinoBertModel
 
-__all__ = ['BertModel']
+__all__ = ['BertModel', 'GroundingDinoBertModel']
