--- projects/GroundingDINO/groundingdino/__init__.py
@@ -1,5 +1,0 @@
-from .dense_heads import GroundingDINOHead
-from .detectors.grounding_dino import GroundingDINO
-from .language_models import GroundingDinoBertModel
-
-__all__ = ['GroundingDINO', 'GroundingDINOHead', 'GroundingDinoBertModel']
