--- projects/GroundingDINO/groundingdino/dense_heads/__init__.py
@@ -1,3 +1,0 @@
-from .grounding_dino_head import GroundingDINOHead
-
-__all__ = ['GroundingDINOHead']
