--- projects/GroundingDINO/groundingdino/language_models/__init__.py
@@ -1,3 +1,0 @@
-from .grounding_dino_bert import GroundingDinoBertModel
-
-__all__ = ['GroundingDinoBertModel']
